Commit 10 Jan - 1
Update Controller Class Diagram ->
added version that uses the Layout as proposed by program (adapted for
printability on A4 page)
It shows a more "downward flow".
Colors: purple for controller, green for database, blue for model and
red for Controller exception.
Attention colors on screen are not "WYSIWYG", prints in different
colours but should just fit on an A4 page.